Move Weight from _core to objectives
