+++ examples/2_Intermediate/stage_two_optimization.py
@@ -25,7 +25,7 @@ import os
 from pathlib import Path
 import numpy as np
 from scipy.optimize import minimize
-from simsopt._core import Weight
+from simsopt.objectives import Weight
 from simsopt.geo import SurfaceRZFourier
 from simsopt.objectives import SquaredFlux
 from simsopt.objectives import QuadraticPenalty
