+++ tests/core/test_derivative.py
@@ -1,8 +1,8 @@
 import unittest
 import numpy as np
-from simsopt._core.optimizable import Optimizable, ScaledOptimizable, OptimizableSum, Weight
+from simsopt._core.optimizable import Optimizable, ScaledOptimizable, OptimizableSum
 from simsopt._core.derivative import Derivative, derivative_dec
-
+from simsopt.objectives.utilities import Weight
 
 class Opt(Optimizable):
     """
